requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
joblib==1.4.2
numpy==2.2.2
pandas==2.2.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.5.0
tzdata==2025.1
urllib3==2.3.0
matplotlib
seaborn